+++ NEWS.md
@@ -1,5 +1,24 @@
-# healthyR.ai 0.0.1
+# healthyR.ai 0.0.2
 
+## Breaking Changes
+None
+
+## New Features
+1. Fix #22 - Add functions:
+  +  `hai_kmeans_mapped_tbl()`
+  +  `hai_kmeans_obj()`
+  +  `hai_kmeans_scree_data_tbl()`
+  +  `hai_kmeans_scree_plt()`
+  +  `hai_kmeans_tidy_tbl()`
+  +  `hai_kmeans_user_item_tbl()`
+2. Fix #29 - Add function `pca_your_recipe()`
+3. Fix #32 - Add function `hai_kmeans_automl()`
+4. Fix #49 - Add function `hai_kmeans_automl_predict()`
+
+## Minor Fixes and Improvements
+None
+
+# healthyR.ai 0.0.1
 
 ## Breaking Changes
 None